requirements.txt
# Main
paramiko
python-dotenv==1.0.1

# UI
streamlit==1.41.1
streamlit_extras
streamlit-modal==0.1.2
streamlit_code_editor
# st-pages==0.5.0
# streamlit-router
# streamlit-navigation-bar
# streamlit-pagination
# streamlit-flow-component==1.2.8
# streamlit_javascript